#36 조회 성능 최적화 1. 기존 어플리케이션 단에서 처리하던 필터링을 DB에서 조회하도록 변경 2. 그로 인한 데이터 불일치 제거
count 쿼리의 최적화가 한번 더 필요할 듯 보임.
no offset으로의 변경도 고려중.

+++ src/main/java/kim/sihwan/daangn/dto/common/Result.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Result<T> {
     private T data;
-    private int totalPage;
+    private long totalPage;
 }
